+++ changelog.md
@@ -1,7 +1,8 @@
 * Updated to 1.21.10
+* Added White Oak, Raw Bamboo & Charred Shelves
+  * Charred shelf are made with charred planks but give 2 shelves instead of 6.
+* Tweaked Raw Bamboo Door, Trapdoor, Grate & Lattice textures
 
 To do :
 * Added Copper Lantern Blocks
 * Added Copper Gate
-* Added White Oak, Raw Bamboo & Charred Shelves
-  * Charred shelf are made with charred planks but give 2 shelves instead of 6.